+++ PhotoApp/Resources/Assets.xcassets/Contents.json
