+++ PhotoApp/Sources/App/AppDelegate.swift
