+++ PhotoApp/Sources/ViewControllers/ViewController.swift
@@ -1,0 +1,29 @@
+import UIKit
+
+class ViewController: UIViewController {
+    //MARK: - Outles
+
+
+
+
+    //MARK: - Lifecycle
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .white
+
+        setupHierarchy()
+        setupLayout()
+    }
+
+
+
+    //MARK: - Setup
+    func setupHierarchy() {
+
+    }
+
+    func setupLayout() {
+
+    }
+}
+
