--- PhotoApp/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  PhotoApp
-//
-//  Created by Никита Шиляев on 14.10.2023.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
